feat: adds layout component for styling about and index
