+++ src/components/layout.js
@@ -1,0 +1,27 @@
+import * as React from 'react';
+import { Link } from 'gatsby';
+
+const pageStyles = {
+  color: "#232129",
+  padding: 96,
+  fontFamily: "-apple-system, Roboto, sans-serif, serif",
+};
+
+const Layout = ({ pageTitle, children }) => {
+  return (
+    <div>
+      <nav>
+        <ul>
+          <li><Link to="/">Home</Link></li>
+          <li><Link to="/about">About</Link></li>
+        </ul>
+      </nav>
+      <main style={pageStyles}>
+        <h1>{pageTitle}</h1>
+        {children}
+      </main>
+    </div>
+  );
+};
+
+export default Layout;
