+++ src/pages/about.js
@@ -1,14 +1,14 @@
-import * as React from 'react'
+import * as React from 'react';
+import Layout from '../components/layout';
 
 const AboutPage = () => {
   return (
-    <main>
-      <h1>About Me</h1>
-      <p>Hi there! I'm Alan. This is the site. Welcome</p>
-    </main>
-  )
-}
+    <Layout pageTitle="About Me">
+      <p>Hi there! I'm Alan. This is the about me page. </p>
+    </Layout>
+  );
+};
 
-export const Head = () => <title>About Me</title>
+export const Head = () => <title>About Me</title>;
 
-export default AboutPage
+export default AboutPage;
